--- app/src/main/java/woowacourse/omok/database/OmokDbHelper2.kt
@@ -1,33 +1,0 @@
-package woowacourse.omok.database
-
-import android.content.Context
-import android.database.sqlite.SQLiteDatabase
-import android.database.sqlite.SQLiteOpenHelper
-
-class OmokDbHelper2(context: Context) : SQLiteOpenHelper(context, DATABASE_NAME, null, DATABASE_VERSION) {
-    override fun onCreate(db: SQLiteDatabase) {
-        db.execSQL(OmokContract2.SQL_CREATE_ENTRIES)
-    }
-
-    override fun onUpgrade(
-        db: SQLiteDatabase,
-        oldVersion: Int,
-        newVersion: Int,
-    ) {
-        db.execSQL(OmokContract2.SQL_DELETE_ENTRIES)
-        onCreate(db)
-    }
-
-    override fun onDowngrade(
-        db: SQLiteDatabase,
-        oldVersion: Int,
-        newVersion: Int,
-    ) {
-        onUpgrade(db, oldVersion, newVersion)
-    }
-
-    companion object {
-        const val DATABASE_VERSION = 1
-        const val DATABASE_NAME = "Omok2.db"
-    }
-}
