--- app/src/main/java/woowacourse/omok/database/OmokEntity2.kt
@@ -1,8 +1,0 @@
-package woowacourse.omok.database
-
-data class OmokEntity2(
-    val x: Int,
-    val y: Int,
-    val color: String,
-    val roomName: String,
-)
